requirements.txt
pytest==7.4.2
selene>=2.0.0rc4
allure-pytest==2.13.2

